OpenStack Glance support
 * package structure has been refactored
 * removed useless files
 * review comments addressed
