+++ acceptance/openstack/imageservice/v2/common.go
@@ -3,11 +3,14 @@
 package v2
 
 import (
+	"bytes"
 	"os"
 	"testing"
 
 	"github.com/rackspace/gophercloud"
+	"github.com/rackspace/gophercloud/acceptance/tools"
 	"github.com/rackspace/gophercloud/openstack"
+	"github.com/rackspace/gophercloud/openstack/imageservice/v2/images"
 	th "github.com/rackspace/gophercloud/testhelper"
 )
 
@@ -44,4 +47,40 @@ func newClient(t *testing.T) *gophercloud.ServiceClient {
 	})
 	th.AssertNoErr(t, err)
 	return c
+}
+
+func createTestImage(t *testing.T, client *gophercloud.ServiceClient) images.Image {
+	//creating image
+	imageName := tools.RandomString("ACCPT", 16)
+	containerFormat := "ami"
+	createResult := images.Create(client, images.CreateOpts{Name: &imageName,
+		ContainerFormat: &containerFormat,
+		DiskFormat:      &containerFormat})
+	th.AssertNoErr(t, createResult.Err)
+	image, err := createResult.Extract()
+	th.AssertNoErr(t, err)
+	t.Logf("Image %v", image)
+
+	//checking status
+	image, err = images.Get(client, image.ID).Extract()
+	th.AssertNoErr(t, err)
+	th.AssertEquals(t, image.Status, images.ImageStatusQueued)
+
+	//uploading image data
+	data := []byte{1, 2, 3, 4, 5, 6, 7, 8, 9}
+	putImageResult := images.PutImageData(client, image.ID, bytes.NewReader(data))
+	th.AssertNoErr(t, putImageResult.Err)
+
+	//checking status
+	image, err = images.Get(client, image.ID).Extract()
+	th.AssertNoErr(t, err)
+	th.AssertEquals(t, image.Status, images.ImageStatusActive)
+	th.AssertEquals(t, *image.SizeBytes, 9)
+	return *image
+}
+
+func deleteImage(t *testing.T, client *gophercloud.ServiceClient, image images.Image) {
+	//deteting image
+	deleteResult := images.Delete(client, image.ID)
+	th.AssertNoErr(t, deleteResult.Err)
 }
