+++ openstack/imageservice/v2/images/urls_test.go
@@ -1,4 +1,4 @@
-package v2
+package images
 
 import (
 	"testing"
